requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pillow==10.1.0
httpx==0.25.2
ollama==0.1.7
python-decouple==3.8
aiohttp==3.9.1
python-magic==0.4.27
structlog==23.2.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1